PLUGIN-75 Marketo Plugin - add some tests.
